fix(state): use integer milliseconds for SQLite busy timeout

+++ internal/state/db.go
@@ -45,7 +45,7 @@ func Open(ctx context.Context, path string) (*DB, error) {
 	}
 	// SQLite's busy handler can outlive context cancellation. Bound its wait by
 	// the remaining command budget as well, rounding up to whole milliseconds.
-	busyMS := (busyTimeout + time.Millisecond - 1) / time.Millisecond
+	busyMS := int64((busyTimeout + time.Millisecond - 1) / time.Millisecond)
 	dsn := u.String() + fmt.Sprintf("?_pragma=foreign_keys(1)&_pragma=busy_timeout(%d)&_pragma=synchronous(FULL)&_txlock=immediate", busyMS)
 	sqldb, err := sql.Open("sqlite", dsn)
 	if err != nil {
